+++ internal/errs/errs.go
@@ -258,3 +258,22 @@ func Unpack(err error) []error {
 	}
 	return result
 }
+
+type ExternalError interface {
+	ExternalError() bool
+}
+
+func IsExternalError(err error) bool {
+	if err == nil {
+		return false
+	}
+
+	for _, err := range Unpack(err) {
+		errExternal, ok := err.(ExternalError)
+		if ok && errExternal.ExternalError() {
+			return true
+		}
+	}
+
+	return false
+}
